server/requirements.txt
aiofiles==23.1.0
fastapi==0.100.0
GitPython==3.1.31
pykwalify==1.8.0
PyYAML==6.0
starlette==0.27.0
litellm>=1.67.0
pydantic>=2.0.0
